% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm.R
\name{rxnorm_class}
\alias{rxnorm_class}
\title{Get drug classes for an RxCUI}
\usage{
rxnorm_class(rxcui, source = "ATC")
}
\arguments{
\item{rxcui}{RxNorm concept ID}

\item{source}{Classification source: "ATC", "VA", "MESH", "MEDRT".
Default "ATC".}
}
\value{
tibble: class_id, class_name, class_type, rxcui, drug_name
}
\description{
Returns ATC, VA, and other classification codes.
}
