% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm.R
\name{rxnorm_context}
\alias{rxnorm_context}
\title{Generate context for the RxNorm package}
\usage{
rxnorm_context()
}
\value{
Character string (invisibly)
}
\description{
Generate context for the RxNorm package
}
