% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm.R
\name{rxnorm_properties}
\alias{rxnorm_properties}
\title{Get drug properties by RxCUI}
\usage{
rxnorm_properties(rxcui)
}
\arguments{
\item{rxcui}{RxNorm concept ID}
}
\value{
tibble: one row with rxcui, name, tty, synonym
}
\description{
Get drug properties by RxCUI
}
