% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm.R
\name{rxnorm_related}
\alias{rxnorm_related}
\title{Get all related concepts for a drug}
\usage{
rxnorm_related(rxcui, tty = NULL)
}
\arguments{
\item{rxcui}{RxNorm concept ID}

\item{tty}{Optional term type filter (e.g. "BN" for brand names,
"SCD" for clinical drugs). Can be a vector.}
}
\value{
tibble: rxcui, name, tty, synonym
}
\description{
Returns all brand names, clinical drugs, dose forms, etc.
}
