% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm.R
\name{rxnorm_spelling}
\alias{rxnorm_spelling}
\title{Get spelling suggestions}
\usage{
rxnorm_spelling(name)
}
\arguments{
\item{name}{Misspelled drug name}
}
\value{
tibble: suggestion (character)
}
\description{
Get spelling suggestions
}
